Rename from tools/ftag to tools/clitag
